Ajout choix de la ville lorsque aucune ville n'a été sélectionnée
